rm sync rabbitmq
