+++ llama_agents/message_queues/base.py
@@ -24,13 +24,15 @@ class BaseChannel(BaseModel, ABC):
     @abstractmethod
     def start_consuming(
         self, process_message: AsyncProcessMessageCallable, message_type: str
-    ) -> Any: ...
+    ) -> Any:
+        ...
 
 
 class MessageProcessor(Protocol):
     """Protocol for a callable that processes messages."""
 
-    def __call__(self, message: QueueMessage, **kwargs: Any) -> None: ...
+    def __call__(self, message: QueueMessage, **kwargs: Any) -> None:
+        ...
 
 
 class PublishCallback(Protocol):
@@ -39,7 +41,8 @@ class PublishCallback(Protocol):
     TODO: Variant for Async Publish Callback.
     """
 
-    def __call__(self, message: QueueMessage, **kwargs: Any) -> None: ...
+    def __call__(self, message: QueueMessage, **kwargs: Any) -> None:
+        ...
 
 
 class BaseMessageQueue(BaseModel, ABC):
